Commit 01-27-2023
Forms/Models: Historical table model create, it is a many-to-many
relationship table with farmland it itself. A description has been added
to farmland model and form to make it easier to find
Input.html: The input template will be adapted to carry out searches
based on the farmlands registered in the db. It will also necessary to
add the reduce and date range fields.