Work with the Calculate Indices form (CIF)
-Forms/Models/Routes: The display of CIF was finished, by the GET and
POST methods. By POST method, now the form will be display the map with
the calculated index with the aditional information provided. JS was
used to get the map_url generated by the GEE API and display it, also it
was used to display the info in the form received by a JSON.
-Methods/Products: Minor changes weres carry out. In Products.py was add
the visualization parameter for the index (vis_index). Switch case
condition for indices was added in Methods.py.